requirements.txt
streamlit
google-generativeai
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
Pillow